requirements.txt
absl-py==1.0.0
attrs==21.3.0
certifi==2021.10.8
cycler==0.11.0
fonttools==4.28.5
future==0.18.2
imbalanced-learn==0.8.0
imblearn==0.0
iso8601==1.0.2
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.21.5
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyparsing==3.0.6
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.7.3
six==1.16.0
virtualenv==16.4.3
wincertstore==0.2
